+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Sentry config is now a scriptable object ([#220](https://github.com/getsentry/sentry-unity/pull/220))
+
+### Fixes
+
+- Setting IsEnvironmentUser to false by default ([#230](https://github.com/getsentry/sentry-unity/pull/230))
+
 ## 0.2.0
 
 ### Features